# swagger/swagger.yml
swagger: "2.0"
info:
  description: Micro service with authentication
  version: "1.0.1"
  title: Micro service
basePath: "/api/v1"
schemes:
  - "https"
  - "http"
paths:
  /user:
    get:
      summary: "Get user details"
      operationId: "getUserDetails"
      security:
      - Bearer: []
      produces:
      - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ServerInfo"
        404:
          description: "not found"
          schema:
            $ref: "#/definitions/ServerInfo"
        500:
          description: "server error"
          schema:
            $ref: "#/definitions/ServerInfo"
  /login:
    post:
      summary: "Login user"
      operationId: "loginUser"
      produces:
      - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ServerInfo"
        500:
          description: "server error"
          schema:
            $ref: "#/definitions/ServerInfo"
  /register:
    post:
      summary: "create new user"
      operationId: "createUser"
      produces:
      - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ServerInfo"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/AuthorizationError"
        500:
          description: "server error"
          schema:
            $ref: "#/definitions/ServerInfo"
  /update:
    put:
      summary: "update user details"
      operationId: "updateUser"
      security:
      - Bearer: []
      produces:
      - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ServerInfo"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/AuthorizationError"
        500:
          description: "server error"
          schema:
            $ref: "#/definitions/ServerInfo"

  /delete:
    delete:
      summary: "delete registered user"
      operationId: "deleteUser"
      security:
      - Bearer: []
      produces:
      - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ServerInfo"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/AuthorizationError"
        500:
          description: "server error"
          schema:
            $ref: "#/definitions/ServerInfo"
  /json/patch:
    patch:
      summary: "patching a json object"
      operationId: "patchJson"
      security:
      - Bearer: []
      produces:
      - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/PatchedObject"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/AuthorizationError"
        500:
          description: "server error"
          schema:
            $ref: "#/definitions/ServerInfo"
  /thumbnail:
    post:
      summary: "resize an image"
      operationId: "resize"
      security:
      - Bearer: []
      produces:
      - "application/x-www-form-urlencoded"
      responses:
        200:
          description: "success"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/AuthorizationError"
        500:
          description: "server error"
          schema:
            $ref: "#/definitions/ServerInfo"
securityDefinitions:
  Bearer:
      type: apiKey
      name: Authorization
      in: header
definitions:
  ServerInfo:
    type: "object"
    properties:
      name:
        type: "string"
      description:
        type: "string"
      version:
        type: "string"
      uptime:
        type: "number"
  PatchedObject:
    type: "object"
    properties:
      name:
        type: "string"
      handle:
        type: "string"
  AuthorizationError:
    type: "object"
    properties:
      error:
        type: "string"
security:
- Bearer: []